backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environs==11.0.0
filelock==3.14.0
idna==3.7
marshmallow==3.21.2
packaging==24.0
pillow==10.4.0
pipenv==2023.12.1
platformdirs==4.2.2
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.1
virtualenv==20.26.2
